+++ CloudHarvestCLI/commands/__init__.py
@@ -1,3 +1,4 @@
+from CloudHarvestCLI.commands.attach.commandset import AttachCommand
 from CloudHarvestCLI.commands.banner.commandset import BannerCommand
 from CloudHarvestCLI.commands.harvest.commandset import HarvestCommand
 from CloudHarvestCLI.commands.plugin.commandset import PluginCommand