+++ CloudHarvestCLI/commands/attach/__init__.py
@@ -1,0 +1,1 @@
+from CloudHarvestCLI.commands.banner.commandset import BannerCommand